Changing the vec3 class to utilize the x86 SSE instructions. This results in a huge speed up!

+++ Sphere.cpp
@@ -5,12 +5,12 @@
 #include "GraphicsObj.h"
 #include "Sphere.h"
 
-Sphere::Sphere(vec3 * position, MATERIAL obj_material, double radius) : GraphicsObj(position, obj_material) {
+Sphere::Sphere(vec3 * position, MATERIAL obj_material, float radius) : GraphicsObj(position, obj_material) {
     this->radius = radius;
 }
 
 bool Sphere::rayIntersection(vec3 origin, vec3 direction) {
-    double test_result;
+    float test_result;
     test_result = pow(vec3::dot(direction, (origin - *position)), 2.0) - pow((origin - *position).magnitude(), 2.0) + pow(radius, 2.0);
     if (test_result < 0)
         return false;
@@ -23,7 +23,7 @@ vec3 Sphere::calculateSurfaceNormal(vec3 surface_pos) {
 }
 
 vec3 Sphere::calculateRayIntersection(vec3 origin, vec3 direction) {
-    double b, root, distance;
+    float b, root, distance;
     root = pow(vec3::dot(direction, (origin - *position)), 2.0) - pow((origin - *position).magnitude(), 2.0) + pow(radius, 2.0);
     b = -(vec3::dot(direction, (origin - *position)));
     distance = std::min(b + root, b - root);
